优化：重构AOP对jfinal 的 Controller和interceptor 注入工作，代码更加容易理解。
