+++ src/main/java/io/jboot/aop/web/ControllerInjectInterceptor.java
@@ -1,4 +1,4 @@
-package io.jboot.web.controller.interceptor;
+package io.jboot.aop.web;
 
 import com.google.inject.Injector;
 import com.jfinal.aop.Interceptor;
@@ -6,13 +6,14 @@ import com.jfinal.aop.Invocation;
 import io.jboot.aop.JbootInjectManager;
 
 /**
- * Guice容器对controller的自动注入
+ * 用于对controller的自动注入
+ * 注意：如果 Controller通过 @Clear 来把此 拦截器给清空，那么此方法（action）注入将会失效
  */
-public class GuiceInterceptor implements Interceptor {
+public class ControllerInjectInterceptor implements Interceptor {
 
     private Injector injector;
 
-    public GuiceInterceptor() {
+    public ControllerInjectInterceptor() {
         injector = JbootInjectManager.me().getInjector();
     }
